Defined COIN_INT_MAX in CoinFinite.hpp in response to Bonmin ticket #13 (part two...)

+++ CoinUtils/src/CoinFinite.hpp
@@ -51,7 +51,7 @@ typedef long long CoinBigIndex;
 #endif
 
 #ifndef COIN_INT_MAX
-#define COIN_INT_MAX ((~((int)0)) >> 1)
+#define COIN_INT_MAX ((int)((~((unsigned int)0)) >> 1))
 #endif
 
 //=============================================================================
